Fix clang-tidy errors.

+++ src/circular_buffer.cpp
@@ -56,7 +56,7 @@ CircularBufferPlacement CircularBuffer::Reserve(size_t num_bytes) noexcept {
 CircularBufferConstPlacement CircularBuffer::Peek(ptrdiff_t index,
                                                   size_t num_bytes) const
     noexcept {
-  assert(static_cast<size_t>(index) + num_bytes <= size());
+  assert(index + num_bytes <= size());
   ptrdiff_t tail = tail_;
   auto first_index = (tail + index) % capacity_;
   auto data1 = data_.get() + first_index;
